Created a theme folder, and moved theme related files there. Updated default contained button styles.
