+++ src/app.tsx
@@ -6,7 +6,7 @@ import CssBaseline from "@material-ui/core/CssBaseline";
 import ApplicationLayout from "templates/application-layout";
 import Routes from "routes";
 import { ThemeProvider } from "@material-ui/styles";
-import theme from "theme";
+import theme from "theme/theme";
 
 const App: React.FC = () => {
     // -----------------------------------------------------------------------------------------
