+++ src/theme/sizes.ts
@@ -1,0 +1,8 @@
+const sizes = {
+    medium: "1rem",
+    large: "2rem",
+    larger: "3rem",
+    largest: "4rem",
+};
+
+export default sizes;
